relocate compactdb module
